Completed Assignment

+++ Binomial.java
@@ -19,8 +19,8 @@ public class Binomial {
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int [] [] memo= new int [n+1][k+1];
+	public static long binomial(int n, int k) {
+		long [] [] memo= new long [n+1][k+1];
 		for (int i=0; i<=n; i++)
 		{
 			for (int j=0; j<=k; j++)
@@ -32,10 +32,10 @@ public class Binomial {
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
 		//// Replace the following statement with your code.
-		return binomial (n,k,memo);
+		return binomial(n,k,memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long [][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
